requirements.txt
aiofiles==23.1.0
aiogram==2.25.2
aiogram-calendar==0.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.6
yarl==1.9.2

